requirements.txt
Flask
python-dateutil
markdown2
gunicorn
